requirements.txt
keras==3.4.1
flask==3.0.3
loguru==0.7.2
Pillow==9.4.0
boto3==1.35.20
tf-keras==2.17.0
Flask-Cors==5.0.0
safetensors==0.4.5
tensorboard==2.17.1
python-dotenv==1.0.1
transformers==4.44.2
opencv-python==4.10.0.84